Rename command_and_control folder to commandAndControl
