+++ app/packs/src/apps/commandAndControl/FocusNovnc.js
