+++ app/packs/src/apps/commandAndControl/NovncStatus.js
