+++ app/packs/src/apps/commandAndControl/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import CnC from 'src/apps/command_and_control/CnC';
+import CnC from 'src/apps/commandAndControl/CnC';
 
 document.addEventListener('DOMContentLoaded', () => {
   const domElement = document.getElementById('CnC');
